Fix the out-of-memory issue in decision tree.
As the number of attributes may exceed 1000, the jump array in the
tree_table was consuming unreasonable amout of memory, even if the number of
actual element in the array was relatively small.  Therefore, we change
the representation of the jump array to two integers to reduce this memory
consumption, and at the same time remove __jump_aggr(), an aggregate
function implemented in PL/pgSQL, to optimize the performance.

MADLIB-381, MADBLI-382