Support EXTENSION in PG 9.1+.
In madpack, if the target platform supports extension, it switches
to create extension files instead of installing with the old way.
EXTENSION in postgres requires two files in its install directory
under $SHAREDIR/extension, and madpack has been avoiding interaction
with platform's directory.  It is still arguable to switch the
install process automatically, maybe let this process be another
option is the better answer.

madlib.control_in is put in madpack directory, per some design
discussion with Florian.  Also this includes a small change in
madpack to show python stack trace on database object creation
error when -v option is specified, which is handy for the diagnosis
of the script problem.