Improve the table existence check method.
The original implementation of MADLIB.__table_exist did not consider schema
for temporary tables.  We reimplement this function in C in order to not be
bothered with corner cases such like not only the temporary table but also
double-quoted names and non-ASCII characters, etc.  Add a test case for the
temporary table issue.  The code is currently under decision_tree, but this
particular function should move to somewhere else more general later.

MADLIB-416