Decision Trees: - C4.5 code refactoring and optimization.
The code changes include two main parts.
1. code reformatting, such as comments refinement, code style consistency, etc.

2. performance improvement.
2.1 continuous attribute optimization: Users can choose 1~n feature
values as the split value. This will not apply to C4.5 for now.
It will be used for RF.

2.2 The operation after found the best attributes will be done in parallel
except for the create child nodes (we sequentially generate the IDs for new
created child nodes).

2.3 Calculate the ACS set feature by feature.

An observed result of this optimization:

KDD dataset (4,898,431 rows), running on 24 segments with 3 hosts

Before: 00:19:30.403121.
After : 00:11:42.736067.