Support Random Forest.
Random Forest (RF) is an ensemble classifier that consists of multiple
decision trees (DTs) and outputs the class that is voted by the majority
of the individual trees.  A prominent advantage of RF is its unexcelled
accurary.  This is the main reason why we support RF in MADlib.  MADlib RF
is based on Leo Breiman and Adele Cutler’s RF algorithm and runs directly
within a database.

Our RF implementation is based on MADlib DT without error based pruning and
reduced error pruning.  But we add a parameter, named max_split_point, to
control how many feature values will be treated as split values.  Using this
approach, in most cases, the total time of training an RF can be decreased
a lot.

MADLIB-419