Support CREATE EXTENSION, take two.
Instead of the previous attempt to support this in madpack time, this time
we try to make the build system to create EXTENSION files so that upcoming
PGXN support can handle it gracefully.  For users who want to install it
as an extension, the process is to download source, configure, and
"make extension-install."  Still the extension is not relocatable, users
can specify schema name on CREATE EXTENSION with SCHEMA option.