Build System: Support CREATE EXTENSION in PG 9.1+.
We add two build targets, extension and extension-install.  The "extension"
target builds and creates extension-related files and "extension-install"
target copies directly to the postgres installed directory without madpack.
Avoiding madpack seems good for the upcoming PGXN support, too.
I tried to find a way to make the extension relocatable, but for now
the extension is not relocatable as a lot of code assumes the fixed
schema name at the time it's installed, but users can specify schema name
on CREATE EXTENSION with SCHEMA option.

MADLIB-316