Support PGXN.
The requirement of PGXN support is, to let users do

 $ pgxn install madlib

to install extension files in postgres install directory.  After that,
users can CREATE EXTENSION in SQL.

The way pgxn client works is download the zip file, unpack it, enter
the directory, configure if any, make install.  To be compatible with
this, we create a Makefile in the top directory when configured, and
let "make install" be "make -C build extension-install".  This becomes
a little surprising behavior to the normal usage out of PGXN, so
the configure creates such Makefile only if the packager indicates,
by creating META.json file in the top directory.  This is automated
with deploy/PGXN/pgxn.sh, which clones an arbitrary git repository and
m4 META.json_in into the cloned directory, then zip it.

We still need to work on the automation of META.json, but for now
we will package as stands and learn how to do it later.

MADLIB-334