Support PGXN.
The requirement of PGXN support is, to let users do

 $ pgxn install madlib

to install extension files in postgres install directory.  After that,
users can CREATE EXTENSION in SQL.

The way pgxn client works is download the zip file, unpack it, enter
the directory, configure if any, make install.  To be compatible with
this, we create a Makefile in the top directory when configured, and
let "make install" be "make -C build extension-install".  This becomes
a little surprising behavior to the normal usage out of PGXN, so
the configure creates such Makefile only if the packager indicates,
by touching PGXN_ZIP file in the top directory.  This is automated
with deploy/pgxn.sh, which clones the latest git and put an empty
PGXN_ZIP file, then zip it.

This commit also contains META.json, which is the only requirement
to upload a module to PGXN distribution site.