Drecision Tree: Fix a bug when growing too deep.
If we use DBL_EPSILON, then the split criterion value(SCV, such as infogain,
gainratio, or gini) will not be treated as zero although they are really zero
(6.66134E-16, 4.44089E-16 or -4.44089E-16). Therefore, we will grow the tree
(create child nodes), since the SCV is not zero (If it's zero, we will treat
it as leaf node). At this point, although we only have two discrete features,
the tree will grow again and again (The SCV is not zero). Therefore, the number
of parent discrete features (the array length) will be exceeded.

Also add a check for dp_ids not to access out of range.  If the length of dp_ids
is greater than or equal to the number of discrete features, it means something
is wrong, such like the the zero-test above is still too precise.

MADLIB-486