Decision Tree: Let dp_ids go down to child nodes when needed.
After double-checking the result tree table, it seems the dp_ids for
the best continuous feature was not correct, since we didn't assign
the dp_ids to the child nodes of a node whose best feature is a continuous
one.  That is, if the best feature for the current leaf node is continuous,
then the new created child nodes should inherit the dp_ids from this leaf
node(the parent node).  This will not change the final result of the tree,
but will increase the performance (more features will be used to calculate
the best feature).

MADLIB-509