Kolmogorov-Smirnov test: Fixed wrong calculation of test statistic
Jira: MADLIB-554

Fixed the following issue:
Whenever, we saw a new row, we computed the distance of the empirical distribution functions at the current position as

    double diff = std::fabs(state.num(0) / state.expectedNum(0)
                    - state.num(1) / state.expectedNum(1));

This could give wrong results in case of ties. After all, we increased state.num(0) and state.num(1) one after the other, so diff could temporarily be larger than it would be at the end of a group of ties.