Support for generating a Greenplum gppkg
Jira: MADLIB-577

The targets are called gppkg_<version> where <version> is of form "4_2". It depends on ${CPACK_PACKAGE_FILE_NAME}.rpm and use the same files (from the _CPack_Packages directory), but it will exclude all files that are specific to different ports.

The initial gppkg version number for MADlib v0.4 will be 1.1 (note that v0.3 was 1.0).