In preparation for design document: Added UseLATEX.cmake to repository
Jira: MADLIB-522

Changes:
- Imported UseLATEX.cmake 1.9.4 from <http://www.cmake.org/Wiki/images/8/80/UseLATEX.cmake>.
Author: Kenneth Moreland <kmorel@sandia.gov>
Copyright 2004 Sandia Corporation.
Under the terms of Contract DE-AC04-94AL85000, there is a non-exclusive license for use of this work by or on behalf of the U.S. Government. Redistribution and use in source and binary forms, with or without modification, are permitted provided that this Notice and any statement of authorship are reproduced on all copies.
- Added several LaTeX files suffixes to .gitignore