Enhance the encoding method
Jira: MADLIB-592

In MADlib 0.4 version, as we encode the training table column by column, the scalability is not good and the encoding time is very long for some datasets with a large number of columns, such as internet_ads (it has about 1600 columns). Now, we encode all the columns of training table in parallel.