Decision Tree: Peformance Improvement on APIs of C4.5 and RF
Jira: MADLIB-608

For C4.5, there is no changes for the training/classification/scoring/
genrules/displaying/cleaning.  For RF, we remove the max_split_points parameter
in the training API, since we don't need it any more as we observed better
performance now.  Another change is that for both C4.5 and RF, we now can use
an encoded table as training table.