C++ AL: Treat MemoryContext as a hint only during memory allocations
Jira: MADLIB-606

We used to respect the request for dbal::AggregateContext when doing memory
allocations, but current PostgreSQL/Greenplum versions do not take any advantage
of transition states allocated in the aggregate context anyway. They still copy
the transition state, whenever the address of the returned state is different
from the address of the input state.

Other uses of the aggregate context (say, auxiliary data/caches for transition
states) are not currently supported by the C++ AL.

MADLIB-606 was eventually the trigger for this change.