Build system: Build pgxn zip file with 'make pgxn'
Jira: MADLIB-623

Added a target for PGXN to the build system. The release number is set in
deploy/PGXN/CMakeLists.txt, the version number is taken from
src/config/Version.yml. This packages the current source repository, so it is
best used on a clean repository.

I also fixed a problem where a directory called "third_party" was created in
${CMAKE_SOURCE_DIR}.