Decision Tree: Insert information of the columns in which all the values are null.
Jira: MADLIB-635

During encoding process, if all the values of a column is null, then we will
remove the column from breakup table in 'ignore' mode.  Therefore, besides
insert the column information in the breakup table to the metatable, we need
to insert those columns in which all the values are null to the meta table.

Also use the quote_literal and quote_ident for the column names.