BugFix: MADLIB-601
quantile_big uses INT/FLOAT to store the number of rows in the table.
If the number of rows in the table exceeds 2^31, the integer will
overflow.  At the same time, cast count(*) to FLOAT may cause issues.

To  fix it, we use INT8 to store the number of rows in the table.