Association Rules: Make the result stable
Jira: MADLIB-461 MADLIB-449

We use the MADLIB_SCHEMA.svec_to_string(svec) as a unique identifier for a svec, so
that the results are stable now.  According to the definition of support and confidence,
the support and confidence of the generated association rules should be >= give values
rather than >.  We also fix that. Furthurmore, we remove the dead code of the
implementation.