Association Rules: Rewrite Apriori for performance improvement
Jira: MADLIB-462 MADLIB-475 MADLIB-638

After studying the current implementation of Apriori, we found the following
issues:

1) The result was unstable, since it used the hash key to unique identify a
   svec.  As we know, two svecs may have the same hash key.  Therefore, we
   will not use the hash key.
2) There were a lot cross joins.  Actually, we can use inner join to do the
   same thing as the cross join did.  The cross join was making the process
   anti-scalable.
3) It was changing the sparse vector in many place, which caused the memory
   issues.  Now, we don't change the svec.  We just construct the svec and
   then do some operations(such as minus, multiply, etc) on that.
4) It used PLPGSQL to implement the algorithm, including aggregate function.

Based on the above analysis, we rewrite it. We use PLPYTHON and C++ AL to
address these.

Although we add code in UDF class for SRF, it is still pretty premature.
It has a lot of spaces to redesign for general purpose.