New k-means: Fixed SQL for compensating for dropped centroids
Jira: MADLIB-522

During the reassignment phase k-means may produce stray centroids. In this case
we sample new centroids using kmeans++. The SQL for this operation had a bug,
which is now fixed. Also, for GP, the SQL needed to be written so that it would
not execute on segment nodes.

I added a unit test for the case where a centroid is dropped (this is easy to
reproduce by simply giving an initial centroid more than once).