New k-means: Query tuning
Jira: MADLIB-678

During development I added an "OFFSET 0" to the main k-means query, which improved
PostgreSQL performance by a factor of 3, because it caused a GroupAggregate to be
replaced by a HashAggregate. On Greenplum, however, the OFFSET 0 hurts performance.
Therefore, the OFFSET 0 is now only added conditionally. In the future, it would
be nice to completely get rid of this implicit hint to the planner.