New k-means: Query tuning
Jira: MADLIB-678

The previous optimization fence "OFFSET 0" was suitable for PostgreSQL only. 
However, an optimization fence is also needed on Greenplum -- with it the innermost
subquery can be sped up by a factor around 2.

An alternative would be moving the innermost subquery that contains the
closest_column calls into a common table expression, but unfortunately that is not
supported by earlier Greenplum versions.

It is unfortunate that we have to use such tricks, but we do not want to forgo a
factor-2 performance improvement.