Revert commit 72b11c1 and provide a more suitable fix for MADLIB-637
Jira: MADLIB-637

The fix in commit 72b11c1 moved logic (i.e., the check whether the number of rows
in a state is 0) back to a place where it not supposed to be (linear.cpp).

I added a unit test that ensures the merge function behaves correctly when there are
no rows on one segment, i.e., when the merge function is called with the aggregate's
initial state in one of its arguments.