Build system: Option LIBSTDCXX_COMPAT for compatibility with an old libstdc++
Jira: MADLIB-263

This commit follows ideas proposed here: http://glandium.org/blog/?p=1901
Unfortunately, MADlib seems to use libstdc++ to a greater extend than Firefox 4.0
did, so we need to do a bit more.

The declarations and definitions in this commit make it possible to build MADlib
with the following versions of gcc (please add to the list), while continuing to
only rely on, e.g., libstdc++.so.6.0.8 (which corresponds to gcc 4.1.2, and
labels GLIBCXX_3.4.8, CXXABI_1.3.1).

Tested with the following versions of gcc:

- gcc 4.4.2
- gcc 4.5.4
- gcc 4.6.2

It should most likely also work with other recent gcc versions.

To use, add the option "-DLIBSTDCXX_COMPAT=40102" when running cmake.