Sketch: Perform strict check of elements only when accessed
Jira: MADLIB-689

In order to fix the security issues of the sketch method described
in MADLIB-630, at the entry point of each UDF, we strictly
check the contents stored in a vararea to make sure that each
access to elements in the vararea is safe. This checking
degenerates the performance.

Because these step functions only access each item at most one time
on each call, in order to avoid performance degeneration, we only
perform the check when the specific element is accessed. The
performance degeneration decreases from 50% at most to 20%.

Other changes:
    - Cleaned up whitespace issues in sketch files