Naive Bayes: Fixed null probability values in test results
Jira: MADLIB-523

This change addresses a issue in previous code that produced NULL
when unobserved feature values were found in test dataset.
For feature values in tests that are have not been observed in the
training data, we compute the probability of the classification
class producing that feature by using Laplace smoothing on prior
data.