Quantile: Suppress assignment from float to integer
Jira: MADLIB-601

quantile_big used BIGINT to store the number of rows. The code
needs to calculate quantile_size by multiplying the total number of
rows with the quantile, whose result type is a FLOAT8. The code
tries to assign this result to a BIGINT directly, and this
operation causes problem. To fix this, we coerce a cast on the
result.