Quantile: Suppress assign float to integer
Jira: MADLIB-601

quantile_big used BIGINT to store the number of rows. The code need to calculate quantile_size by multiplying the total number of rows with the quantile, whose result type is a FLOAT8. The code tries to assign this result to a BIGINT directly, and this operation causes problem.

To fix this, we make a coerce on the result.