Multilogistic: Apply input validation for depvar, numcategories
Jira: MADLIB-673, 733

Dependent variable is of the form '1 of K values'. Hence, we only
support integer values for depvar. If the 'depvar' column in the
'source' table is not of type INTEGER then an exception is raised.
The 'numcategories' is redundant since we only develop as many
models as distinct categories present in the training data. If the
user specifies a 'numcategories' parameter that is greater than the
actual number of categories then it could lead to overflow/underflow errors, hence a warning is presented in such a
scenario.

Other changes:
    - Trailing whitespace removed
    - Other code cleanup in comments