Naive Bayes: Revert changes for 'null probabilities' fix
Jira: MADLIB-745, MADLIB-746
This reverts commit fdd4ca0, 238d4dd, and d7cd128 due
to out-of-memory and performance issues introduced by array_agg
operator on big datasets. We switch to the previous code that uses
'join' instead of 'array_agg'.
For future work, though, we still need to find a fix for
NULL probability values in test results when the test case contains
a value not present in the training set (MADLIB-523). We also need
to check for correctness of results (MADLIB-679).