NB Documentation: Add explanation for classification differences
Jira: MADLIB-679

(1) Probabilities computed on PostgreSQL and Greenplum have small
differences due to floating point computation. If a data point has
equal (or very close) probabilities for two classes, this point
might be classified into diferent classes on PostgreSQL and
Greenplum. This leads to the classifications with a small
differences on PostgreSQL and Greenplum for some data sets, but
does not affect the quality of the results.

(2) The current implementation of Naive Bayes classification is
only suitable for discontinuous attributes. One can still feed
continuous data into the naive Bayes classification function. No
warning is raised, but the result may not be accurate.