NB Documentation: Add further clarification for results
Jira: MADLIB-679

When the probabilities for two classes are equal and the maximum among all classes, the classification result is an array of these two classes, the order being random.

One can still provide floating point data to the naive Bayes
classification function. Floating point numbers, however, are
currently interpreted as categorial data. No warning is raised if
the attribute data is provided as continuous, and the results are
undefined.