NB Documentation: Add more explanation for classification
Jira: MADlib-679

When the probabilities for the two classes are exactly equal, the result is
an array of these two classes, but the order of the two classes is random.

One can still provide continuous (floating point) data to the naive Bayes
classification function. Floating point numbers can be used as categorial data.