K-means: Fix dist_angle metric to return values when norm is 0
Jira: Madlib-768

When distance metric 'madlib.dist_angle' is used, which is defined as
the angle between two vectors, Inf is obtained when either of the two
norms is 0. The angle is not defined mathematically. Just let the
metric return \pi under such circumstances to avoid error. This should
not affect the quality of the result.