Abstraction Layer: Allow UDFs to access the protected member fcinfo and mSysInfo of class AnyTypet
Jira: MADLIB-775
The UDFs cannot access the protected members fcinfo and mSysInfo of class
AnyType, although the class UDF is declared as friend of class AnyType (Note
that the friendship cannot be inherited in C++). The reasons for a UDF to
access fcinfo include judging whether it is called in an aggregate state or
window state. The reasons for a UDF to access mSysInfo include using
mSysInfo->user_fctx to keep some internal states.