Abstraction Layer: Incompatibility of construct_array and construct_md_array between Greenplum and PostgreSQL
Jira: MADLIB-776
The construct_array and construct_md_array have different behaviors in
Greenplum and PostgreSQL. When the first parameter is set to NULL in both
functions, Greenplum can work well, but PostgreSQL will crash the system. The
reason is that PostgreSQL does not check for null pointer and always calls
CopyArrayEls even the pointer is null.