Naive Bayes: Skip NULL values (not rows) from training
Jira: MADLIB-614, 749

NULL values in attribute values cannot be directly used for training
the model. Currently rows with NULL values are completely ignored.
This becomes a problem in datasets that have many rows with at least
one NULL value. We can currently ignore just the specific attribute
which is NULL instead of ignoring the whole row. This should be
changed to better NULL handling, similar to the feature in R.