Cross Validation: Create a cross validation module
JIRA: PIVOTAL-40975391

Created a new module for applying cross validation on regression
methods. The current interface is setup to work with linear and
logistic regression, as well as ridge, lasso, and elastic net
regression techniques that are upcoming. All other modules can be
supported by adding wrapper functions to work with the cross
validation interface