Utilities: Use madlib array_agg for GPDB < 4.1 and PG < 9.0
Updated utilities function is_col_exists to use the
madlib_schema.array_agg when the version of GPDB is 4.0 and version
of PG is 8.4. For this change, the function takes an extra parameter,
the madlib_schema name. This change was reflected in logistic
regression and cross validation where is_col_exists was being used.