New module: Upgrade
Added new functionality to upgrade MADlib deployment on a database without
removing the previous installation. This is currently only supported for version
0.5 and greater. We don't yet support upgrade in Postgresql systems since
MADlib installs a Bool-to-Text cast in Greenplum and the upgrade tries to drop
the cast. This leads to a problem in Postgresql and should be handled as an
exception in future releases.