Matrix Operations: Fix order in Column to array operation
The column to array functionality to build a matrix from a table
extracted the list of column names from information schema ordered  by
the ordinal position. This order, however, was lost when the column
names were stored in a 'set' instead of a 'list'. This commit fixes
the problem ensuring that the order of the columns is maintained
while building the matrix.