Specifically use /bin/bash in madpack
In general /bin/sh is equivalent to /bin/bash, but not on ubuntu where /bin/sh is actually /bin/dash.  This matters because the madpack script uses `pushd` and `popd` which are bash-specific enhancements; on ubuntu this script causes errors like:

    /usr/local/madlib/bin/madpack: pushd: not found
    /usr/local/madlib/bin/madpack: popd: not found

The fix is to specifically use bash.  For more info:

    http://askubuntu.com/questions/141928/what-is-difference-between-bin-sh-and-bin-bash