Bug fix: cannot build gppkg on Linux with newer rpmbuild
Pivotal Tracker #66443692

Newer rpmbuild uses a default build root, which is different from the file
directory that we use and 'make gppkg' fails. The solution is to create
a link in the new default build root pointing to the folder that has all
the necessary files. See the comments in the changed file for the
detailed explanation.