Elastic Net: Use loglikelihood to test for convergence
Pivotal Tracker: 62078786
JIRA: MADLIB-830

For an elastic net problem, since the number of features is more than the number
of data points, the coefficient result could be unstable in some cases. For
small datasets, the coefficients tend to oscillate around an optimal, leading to
longer convergence time. Using loglikelihood to test for convergence leads to
quicker convergence.

In some cases the convergence can be premature. Hence the default threshold for
convergence test has been reduced from 1e-4 to 1e-6.