Decision tree: Design document for surrogate splits
Pivotal Tracker: 79902020

Surrogate splits are secondary splits for each node that can be used to
redirect a tuple through the tree when the tuple has a NULL value for
the feature used in the node split. This commit includes design
considerations for adding surrogate splits to the current decision tree
design.