Decision Tree: Add support for NULL indep using surrogate splits
Pivotal Tracker: 75480136

Surrogate splits are used to compute the split for a node when the value
of the feature in primary split is NULL. The surrogate splits are
computed by going through the data after a node is trained and
computing the number of rows that are split to the same side when
compared between the primary split and every other split.
The top such splits are chosen as surrogates for the primary split.

Other bug fixes:
  - The tree depth in output table during grouping reflected the depth
    of the largest tree. This has been fixed to give the depth of each
    tree.
  - The tree depth is now being updated after pruning to reflect the
    actual depth.
  - The rebinding of backup tree did not correctly allocate memory for a
    classification tree, since the 'is_regression' parameter was not
    being passed into rebind.