Random forest: Add surrogates to support NULL values
Pivotal tracker: 80472868

Surrogates are used in decision trees to handle rows that have NULL
value for some features. If such a feature is used as a split variable
in a node, the surrogates for the feature would be used to split the
row. A surrogate parameter is added in random forest and passed to
decision tree. A function to display the surrogates of a particular tree
in the forest is also provided.