Random forest: Create duplicate rows in bootstrap sample
Pivotal tracker: 82584376

Additional author: Feng, Xixuan (Aaron) <xfeng@pivotal.io>

Random forest used weights as an indicator of duplicate rows
in the bootstrap sample to decision tree. This led to an incorrect
count of rows in the tree display and surrogate calculation. This has
been fixed by creating a view that duplicates the rows instead of
incrementing the weight for the row.

We also add a bug fix to ensure the majority branch count in the
surrogate display corresponds to the majority count used during
surrogate training. After surrogate training, the majority count
increases since it includes the rows with NULL values directed by the
surrogate splits.