[Performance] Improve decision tree CV performance on HAWQ
Pivotal tracker: 83515716

Author: Preethi, Hai

Greatly improved the decision tree cross-validation performance, but it
is still quite slow. Here reducing the number of locks used and
minimizing the changes to catalog improves a lot. On HAWQ, most of the
time is spent after the computation at the COMMIT stage.

Other changes: exposed random forest help message function and updated
the help messages.