Multiple: Add distribution policy for all create statements
Pivotal Tracker: 84236616

ORCA always uses DISTRIBUTED RANDOMLY for a "CREATE TABLE AS" if no
distribution policy is provided. For multiple modules we create tables
without providing a distribution policy while creating a primary key in
the table. This leads to an incompatibility between random distribution
and primary key constraint. We solve this problem by adding a
distribution policy that is same as the primary key.