Validation: Add option in table_exists to search only 1st schema in path
Pivotal Tracker: #87646110

Details:
The table_exists function in validate_args checks for the table in all
schemas in the search path. For output tables, we only want to check in
the current schema. We add a boolean as a flag to
differentiate the two situations and change all calls when checking for
output table.

The commit also includes updates to all modules that use table_exists,
with the flag set to True when the output table is validated.

Others:
Add online help message to coxph_predict function
Add drop functions and change lists for coxph functions for HAWQ
reinstall and upgrade, missed in v1.6