Matrix ops: Support id of any type that can cast to int
Pivotal tracker: 96013656

Changes:
    - Add casts in matrix ops, SVD, PCA to allow row_id/col_id of
      any type that can be cast to int.
    - Improve performance in dense multiplication by transposing the
      second matrix only if required.

After this change, the row_id for dense matrix and row_id, col_id for
sparse matrix can be of any type. For some operations (like
matrix_densify), the id is cast to be of type integer. Further, in some
cases, we check the size of the row/col dimension to ensure it is
within the range for integer.