Matrix ops: Add output format specifier
Pivotal tracker: 98059402

Additional author: Yu Yang <myang@pivotal.io>

An output format specifier (fmt) has been added to the 'out_args' argument
that allows the user to specify the exact format of the output matrix.
Previously, the output matrix format was inferred from the input
matrices. This is still the default behavior if 'fmt' is not provided.
The format specifier is ignored in the matrix_densify and
matrix_sparsify functions, since the objective of the functions is to
output matrices in a specific format.