Path: Match a pattern in a subset of partition
JIRA: MADLIB-916

This commit adds the functionality to match a subset of a partition
instead of just the complete partition. This is performed by
array-agging the symbols and then finding the position of a match within
the agg. This position is then compared to an array of ids to get the
actual rows that correlate to this match.

This approach will only work if each symbol is a single character. Since
we allow the user to set symbols to be arbitrary strings, the
user-supplied symbol is replaced with a single character (in the row
match and in the pattern expression).

Other: Parent commit (30e9286) closes #9