Utilities: Revert 60a07eb + use unquoted table names
The previous commit (60a07eb) used pg_attribute instead of
information_schema to get column names and types. This led
to different names in the types causing issues in rest of the code. It's
better to use information_schema since it gives general (non-Postgres
specific) results, but it requires unquoted table and schema names as
input.