Correlation: Return columns sorted in ordinal position
JIRA: MADLIB-941

Couple of minor issues fixed here:
1. get_cols_and_types utility function returned columns
in an arbitrary order due to dictionary creation at the end. This has
been fixed by returning a list of tuples. An OrderedDict would be the
best choice here but some platforms are still on Python 2.6
2. Multiple modules that depended on this function had to be updated
either creating the dict in the calling function or using the list
instead of a dict.