Elastic Net: Check only if features are numeric
JIRA: MADLIB-952

Columns were being checked to ensure every column is of the same numeric
type. While giving an error for non-numeric types is correct, there is
no need to enforce them to be of same numeric type, as the columns are
cast to float8[] inside the function. The input analyzer has been
changed to relax this condition.

This closes #12.