SVM: Add polynomial kernel transformation
JIRA: MADLIB-938

The features are transformed using Kar and Karnick's appoach, which
applies to any positive definite kernel where the kernel function
admits a Maclaurin expansion. This commit implements the special
case when the kernel function is polynomial.

As part of this work, we also add support for Bernoulli distribution
to matrix_random.