Random Forest: Fix filtered feature related bug
JIRA: MADLIB-928

Additional Author: Nandish Jayaram <njayaram@pivotal.io>

Random forest filters out a feature if it has the same value for every
row. If grouping is enabled, this filer is applied per group. However
_get_bins_grps function did not consider a case where different groups
have different feature sets. The commit fixes this issue, adds a test
for install-check to replicate the use case and fixes a typo in one of
the related functions.