Feature: Sessionize funtion
JIRA: MADLIB-909

This contains the initial implementation of the sessionize function.
The current input parameters are the input and output table names,
along with a partition expression, the time_stamp column name and
the time_out period to consider for sessionization. The implementation
uses the window function to perform sessionization.
This commit also contains the install check file for sessionization.

Closes #44