Sessionize: Allow NULL input for partition expression
JIRA: MADLIB-1001

Partition expression can now be set to NULL or '' to run sessionization on the
complete table. This is in line with how path handles partition_expr.
Additionally, if output_expr has NULL or '' as its value, it is set to '*',
which is its default value. Finally, the create_view param is also set to its
default value if input is NULL. This commit also contains some formatting
changes to the online help.

Closes #58