Build: Fix version used in rpm installation
During MADlib installation a directory is created in `Versions` that
contains the current version. A soft link called `Current` is created
in the root for madpack to link to this directory.

Versions that have a hyphen were invalid for gppkg and hence were
updated to replace hyphen with an underscore. This led to discrepancy
with the softlink. This commit fixes that bug by ensuring the softlink
links to the right place.