Association Rules: Add rule counts and limit itemset size feature
JIRA: MADLIB-1044, MADLIB-1031

- The output column currently lists support, along with other columns.
The output table now contains count too, which is essentially
support*number_of_transactions for any rule. This could have been
obtained by post-processing the original output table content too,
but listing it explicitly might be useful.
- Add a new optional parameter named max_itemset_size that determines
the maximum size of frequent itemsets that are used for generating
association rules. Must be 2 or more, and the default behavior is to
generate itemsets of all sizes. This parameter can be used to reduce
run time for data sets where itemset size is large.
- This commit also includes improved online and user documentation
with examples of usage.

Closes #87